requirements.txt
openai==1.30.1
python-dotenv==1.0.1
numpy==1.26.4
psutil==5.9.8
tqdm==4.66.4
colorama==0.4.6
scipy==1.13.0
chromadb>=0.4.0
tiktoken>=0.7.0
watchdog==4.0.0
pypdf==4.2.0
